Modified Search Unit Tests
