+++ UnmaskedTests/MockHeroSearchApiClient.swift
